+++ tests/all/main.rs
@@ -1,0 +1,11 @@
+extern crate copy_dir;
+extern crate failure;
+#[macro_use]
+extern crate serde_derive;
+extern crate serde_json;
+extern crate tempdir;
+extern crate wasm_pack;
+
+mod manifest;
+mod readme;
+mod utils;
